+++ CHANGELOG.md
@@ -53,6 +53,7 @@
     - print mask is now visible for group layer
     - zIndex and transparency from the group layer are also adopted in printing
 - Filter: onDownloadProgress function will be commented out in axios in filter.
+- LayerInformation: The entry `show_doc_url` in the metadata is now used again to retrieve further information.
 
 ---
 ## 2024-06-27 v3.0.0